bago matulog

+++ NiisMenu/UpdateMenu.xaml.cs
@@ -4,69 +4,38 @@ using Microsoft.Maui.Storage;
 using NiisMenu.Builder;
 using Plugin.CloudFirestore;
 
-namespace NiisMenu;
-
-public partial class UpdateMenu : ContentPage
+namespace NiisMenu
 {
-	public UpdateMenu(Menu selectedMenu)
-	{
-		InitializeComponent();
-        NameEntry.Text = selectedMenu.Name;
-        PriceEntry.Text = selectedMenu.Price.ToString();
-        CategoryPicker.SelectedItem = selectedMenu.Category;
-        AvailableSwitch.IsToggled = selectedMenu.IsAvailable;
-        
-	}
-   
-    private async void OnUpdateItemClicked(object sender, EventArgs e)
+    public partial class UpdateMenu : ContentPage
     {
-        try
+        private string existingImageUrl;
+
+        public UpdateMenu(Menu selectedMenu)
         {
-            Menu Newmenu = new Menu
-            {
-                Name = NameEntry.Text,
-                Price = Convert.ToInt32(PriceEntry.Text),
-                Category = CategoryPicker.SelectedItem.ToString(),
-                IsAvailable = AvailableSwitch.IsToggled,
-                Date = DateOnly.FromDateTime(DateTime.Now).ToString(),
-                Time = TimeOnly.FromDateTime(DateTime.Now).ToString()
-            };
+            InitializeComponent();
+            NameEntry.Text = selectedMenu.Name;
+            PriceEntry.Text = selectedMenu.Price.ToString();
+            CategoryPicker.SelectedItem = selectedMenu.Category;
+            AvailableSwitch.IsToggled = selectedMenu.IsAvailable;
 
-
-            var querySnapshot = await CrossCloudFirestore.Current
-                .Instance
-                .Collection("Menu")
-                .WhereEqualsTo("Name", NameEntry.Text)
-                .GetAsync();
-
-            foreach (var documentSnapshot in querySnapshot.Documents)
-            {
-                var documentId = documentSnapshot.Id;
-
-
-                await CrossCloudFirestore.Current
-                    .Instance
-                    .Collection("Menu")
-                    .Document(documentId)
-                    .UpdateAsync(Newmenu);
-            }
-            MessagingCenter.Send(this, "DataUpdated");
-            await DisplayAlert("Success", "Menu Updated", "OK");
-            await Navigation.PopAsync();
-        }
-        catch (Exception ex)
-        {
-            await DisplayAlert("Error", ex.Message, "OK");
+            existingImageUrl = selectedMenu.ImageUrl; // Assign selectedMenu's ImageUrl to existingImageUrl
         }
 
-    }
-    private async void OnDeleteItemClicked(object sender, EventArgs e)
-    {
-        try
+        private async void OnUpdateItemClicked(object sender, EventArgs e)
         {
-            var confirmDelete = await DisplayAlert("Confirm", "Are you sure you want to delete this menu item?", "Yes", "No");
-            if (confirmDelete)
+            try
             {
+                Menu Newmenu = new Menu
+                {
+                    Name = NameEntry.Text,
+                    Price = Convert.ToInt32(PriceEntry.Text),
+                    Category = CategoryPicker.SelectedItem.ToString(),
+                    IsAvailable = AvailableSwitch.IsToggled,
+                    Date = DateOnly.FromDateTime(DateTime.Now).ToString(),
+                    Time = TimeOnly.FromDateTime(DateTime.Now).ToString(),
+                    ImageUrl = existingImageUrl // Use existingImageUrl here
+                };
+
                 var querySnapshot = await CrossCloudFirestore.Current
                     .Instance
                     .Collection("Menu")
@@ -81,20 +50,51 @@ public partial class UpdateMenu : ContentPage
                         .Instance
                         .Collection("Menu")
                         .Document(documentId)
-                        .DeleteAsync();
+                        .UpdateAsync(Newmenu);
                 }
-
                 MessagingCenter.Send(this, "DataUpdated");
-                await DisplayAlert("Success", "Menu deleted successfully", "OK");
+                await DisplayAlert("Success", "Menu Updated", "OK");
                 await Navigation.PopAsync();
             }
+            catch (Exception ex)
+            {
+                await DisplayAlert("Error", ex.Message, "OK");
+            }
         }
-        catch (Exception ex)
+
+        private async void OnDeleteItemClicked(object sender, EventArgs e)
         {
-            await DisplayAlert("Error", ex.Message, "OK");
+            try
+            {
+                var confirmDelete = await DisplayAlert("Confirm", "Are you sure you want to delete this menu item?", "Yes", "No");
+                if (confirmDelete)
+                {
+                    var querySnapshot = await CrossCloudFirestore.Current
+                        .Instance
+                        .Collection("Menu")
+                        .WhereEqualsTo("Name", NameEntry.Text)
+                        .GetAsync();
+
+                    foreach (var documentSnapshot in querySnapshot.Documents)
+                    {
+                        var documentId = documentSnapshot.Id;
+
+                        await CrossCloudFirestore.Current
+                            .Instance
+                            .Collection("Menu")
+                            .Document(documentId)
+                            .DeleteAsync();
+                    }
+
+                    MessagingCenter.Send(this, "DataUpdated");
+                    await DisplayAlert("Success", "Menu deleted successfully", "OK");
+                    await Navigation.PopAsync();
+                }
+            }
+            catch (Exception ex)
+            {
+                await DisplayAlert("Error", ex.Message, "OK");
+            }
         }
     }
-
-
-
 }
